Phase 1: The Walking Skeleton
This is the first simple phase of this project. It downloads the walking map of Hoboken, NJ and finds the shortest path between the Hoboken Terminal and Stevens Institute of Technology. To accomplish this, graph_builder.py downloads the graph and stores it as an actual graph, where each node is an intersection (or sharp change in direction), and each edge is a walking path, and pathfinder.py finds the shortest path based on input starting coordinates and ending coordinates. pathfinder.py first finds the nodes in the graph that are closest to the given coordinates, and then runs a shortest path algorithm to find the shortest path between the starting and ending node (using networkx.shortest_path for now). The script run_phase_1.py runs both the functions in graph_builder.py and pathfinder.py, and allows customization for what city, starting coordinates, and ending coordinates, and it outputs information about the path that was found (how many intersections, what intersections (nodes) were visited, and the total walking distance in meters).

Future planned improvements: Add the frontend interactive interface to both run the script and also view the graph in a way that makes sense to the user with a clean UI. Also, obviously add more than just walking (driving, transit, etc.) and add the functionality to consider the user's preferences. For the walking stage only, preferences might include walking distance, intersections, etc. (for example, some people might prefer to walk a little longer if it means they have to go through less intersections).